+++ gamemodes/terrortown/entities/weapons/weapon_ttt_knife.lua
@@ -48,6 +48,7 @@ SWEP.Kind = WEAPON_EQUIP
 SWEP.CanBuy = {ROLE_TRAITOR} -- only traitors can buy
 SWEP.LimitedStock = true -- only buyable once
 SWEP.WeaponID = AMMO_KNIFE
+SWEP.StockEquipment = true
 
 SWEP.IsSilent = true