+++ gamemodes/terrortown/entities/weapons/weapon_ttt_m16.lua
@@ -19,6 +19,7 @@ SWEP.Base = "weapon_tttbase"
 
 SWEP.Kind = WEAPON_HEAVY
 SWEP.WeaponID = AMMO_M16
+SWEP.StockEquipment = true
 SWEP.spawnType = WEAPON_TYPE_HEAVY
 
 SWEP.Primary.Delay = 0.19