+++ gamemodes/terrortown/entities/weapons/weapon_zm_mac10.lua
@@ -19,6 +19,7 @@ SWEP.Base = "weapon_tttbase"
 
 SWEP.Kind = WEAPON_HEAVY
 SWEP.WeaponID = AMMO_MAC10
+SWEP.StockEquipment = true
 SWEP.spawnType = WEAPON_TYPE_HEAVY
 
 SWEP.Primary.Damage = 12