+++ gamemodes/terrortown/entities/weapons/weapon_zm_sledge.lua
@@ -22,6 +22,7 @@ SWEP.AutoSpawnable = true
 
 SWEP.Kind = WEAPON_HEAVY
 SWEP.WeaponID = AMMO_M249
+SWEP.StockEquipment = true
 SWEP.spawnType = WEAPON_TYPE_HEAVY
 
 SWEP.Primary.Damage = 7