+++ lua/terrortown/entities/items/item_ttt_disguiser.lua
@@ -12,6 +12,7 @@ end
 
 ITEM.CanBuy = {ROLE_TRAITOR}
 ITEM.oldId = EQUIP_DISGUISE or 4
+ITEM.StockEquipment = true
 
 if CLIENT then
 	local trans