+++ lua/terrortown/entities/items/item_ttt_noenergydmg.lua
@@ -11,6 +11,7 @@ ITEM.CanBuy = {ROLE_TRAITOR, ROLE_DETECTIVE}
 
 ITEM.hud = Material("vgui/ttt/perks/hud_noenergydmg.png")
 ITEM.material = "vgui/ttt/icon_noenergydmg"
+ITEM.StockEquipment = true
 
 if SERVER then
 	hook.Add("EntityTakeDamage", "TTT2NoEnergyDmg", function(target, dmginfo)