+++ lua/terrortown/entities/items/item_ttt_radar/shared.lua
@@ -11,6 +11,7 @@ ITEM.EquipMenuData = {
 ITEM.material = "vgui/ttt/icon_radar"
 ITEM.CanBuy = {ROLE_TRAITOR, ROLE_DETECTIVE}
 ITEM.oldId = EQUIP_RADAR or 2
+ITEM.StockEquipment = true
 
 ---
 -- @ignore